requirements.txt
aniso8601==2.0.0
astroid==1.6.3
autopep8==1.3.5
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
grpcio==1.11.0
grpcio-tools==1.11.0
idna==2.6
isort==4.3.4
itsdangerous==0.24
jieba==0.39
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.14.1
protobuf==3.5.2.post1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycodestyle==2.4.0
pylint==1.8.4
pytz==2018.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.6
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
